Add existence checks before registering listeners to runtime.onStartup and onInstalled
Firefox currently does not support these events (http://arewewebextensionsyet.com/#runtime-events),
and trying to register listeners causes errors saying something like `onStartup/onInstalled is undefined`.
Therefore we have to check if these entry points are available or not.
It doesn't seem to have significant effects that badge.updateForActiveTab() is not called
during the startup and after the installing on Firefox.