Migrate React components to ES6 class style
We use ES6 class style (React.Component) for every React component
instead of React.createClass or stateless function for being consistent.

It is so annoying to switch between definition styles every time
depending on what a component does.

To define propTypes in class definition, it introduces
babel-plugin-transform-class-properties plugin as well as babel-eslint.