feat: Support new options from no-mix-dearu-desumasu
This translates options from textlint-rule-no-mix-dearu-desumasu.

To allow an empty string value in enum, we set `allowEmptyOption` to
Selectize, but it is not working properly for the current version of
Selectize. So we take a monkey-patch for it.

Also, `required_by_default` options is set to json-editor to prevent a
problem that option editors for new options are removed when there is
an option value loaded from the stroage.

Translation is refactored as titles in schema get translated instead of
labels after rendered.